feat: http checks + assertions, refactor
