+++ deps.ts
@@ -1,0 +1,12 @@
+export * as ansi from "https://deno.land/std@0.166.0/fmt/colors.ts";
+export { $ } from "https://deno.land/x/dax@0.15.0/mod.ts";
+export { Buffer } from "https://deno.land/std@0.166.0/io/buffer.ts";
+
+// Node compatibility
+export {
+  Buffer as NodeBuffer,
+} from "https://deno.land/std@0.166.0/node/buffer.ts";
+export { Stream as NodeStream } from "https://deno.land/std@0.166.0/node/stream.ts";
+export { Readable as NodeReadable } from "https://deno.land/std@0.166.0/node/stream.ts";
+export { Writable as NodeWritable } from "https://deno.land/std@0.166.0/node/stream.ts";
+export { EventEmitter as NodeEventEmitter } from "https://deno.land/std@0.166.0/node/events.ts";
