+++ version.ts
@@ -1,4 +1,4 @@
-import * as ansi from "https://deno.land/std@0.151.0/fmt/colors.ts";
+import { $, ansi } from "./deps.ts";
 
 /** `VERSION` managed by https://deno.land/x/publish */
 export const VERSION = "0.0.1";
@@ -6,43 +6,43 @@ export const MODULE = "dis";
 
 /** `prepublish` will be invoked before publish */
 export async function prepublish(version: string) {
-  for (
-    const filename of [
-      "README.md",
-      "mod.ts",
-      "types.d.ts",
-    ]
-  ) {
-    await bump(filename, version);
+  for await (const file of $.fs.expandGlob("./*.{md,ts}")) {
+    if (file.isFile) await bump(file.path, version);
   }
 }
 
 /** `postpublish` will be invoked after publish */
 export function postpublish(version: string) {
   console.log(
-    ansi.bgGreen(" SUCCESS "),
-    ` ✓ published ${ansi.bold(ansi.underline(MODULE + "@" + version))}`,
+    ansi.brightGreen(
+      ` ✓ published ${ansi.bold(ansi.underline(`dis@${version}`))}`,
+    ),
   );
 }
 
-async function bump(filename: string, version: string) {
-  try {
-    const module_regex = new RegExp(
-      `(?<=[/"'\s])(${MODULE})[@]([{]{1,2}VERSION[}]{1,2}|\$VERSION|[^/"'\s]+)(?=[/"'\s])`,
-      "ig",
-    );
-    const content = await Deno.readTextFile(filename);
+export async function bump(path: string, version: string, prev = VERSION) {
+  const filename = $.path.basename(path);
+  const _dirname = $.path.dirname(path);
 
-    await Deno.writeTextFile(
-      filename,
-      content.replace(module_regex, `$1@${version}`),
+  try {
+    const TAG = "VERSION";
+    const PLACEHOLDERS = `\{${TAG}\}|\{\{${TAG}\}\}|\$${TAG}|${VERSION}`;
+    const SPECIFIER_RE = new RegExp(
+      `(?<=(?:^|[/"'\s])${MODULE}[@])(${PLACEHOLDERS}|[^/"'\s]+)(?=$|[/"'\s])`,
+      "mig",
     );
-  } catch (e) {
+    let content = await Deno.readTextFile(path);
+
+    if (SPECIFIER_RE.test(content)) {
+      content = content.replaceAll(SPECIFIER_RE, version),
+        await Deno.writeTextFile(path, content).catch(console.error);
+    }
+  } catch (error) {
     console.error(
       ansi.bgRed(" FAILED "),
-      `⚠︎ could not update ${ansi.underline(ansi.italic(filename))} to ${
-        ansi.bold(version)
-      }!\n\n${e}`,
+      `⚠︎ Unable to bump ${ansi.underline(ansi.red(filename))} from ${
+        ansi.underline(prev)
+      } to ${ansi.bold(version)}!\n\n${error}`,
     );
   }
 }
